# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dragon Fruit Farm Store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  unit: string;
  imageUrl: string;
  badge?: string | null;
  inStock: boolean;
  featured: boolean;
  createdAt: string;
}

export interface CreateProduct {
  name: string;
  description: string;
  price: number;
  unit: string;
  imageUrl: string;
  badge?: string | null;
  inStock: boolean;
  featured: boolean;
}

export interface Review {
  id: number;
  customerName: string;
  rating: number;
  comment: string;
  avatarUrl?: string | null;
  createdAt: string;
}

export interface GalleryImage {
  id: number;
  imageUrl: string;
  caption: string;
  category: string;
}

export interface StoreStats {
  totalProducts: number;
  happyCustomers: number;
  yearsOfFarming: number;
  citiesDelivered: number;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}