# lib/api-zod/src/generated/types/createProduct.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dragon Fruit Farm Store API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProduct {
  name: string;
  description: string;
  price: number;
  unit: string;
  imageUrl: string;
  badge?: string | null;
  inStock: boolean;
  featured: boolean;
}