# lib/api-zod/src/generated/types/galleryImage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dragon Fruit Farm Store API
 * OpenAPI spec version: 0.1.0
 */

export interface GalleryImage {
  id: number;
  imageUrl: string;
  caption: string;
  category: string;
}