# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dragon Fruit Farm Store API
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: number;
  customerName: string;
  rating: number;
  comment: string;
  avatarUrl?: string | null;
  createdAt: string;
}