# lib/api-zod/src/generated/types/storeStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dragon Fruit Farm Store API
 * OpenAPI spec version: 0.1.0
 */

export interface StoreStats {
  totalProducts: number;
  happyCustomers: number;
  yearsOfFarming: number;
  citiesDelivered: number;
}